Move importlauncher to file + sync hide logic
